refactor: clean up and reorder
