+++ libs/gui/components/src/lib/components/date-time-calendar.ts
@@ -154,8 +154,6 @@ export class GuiDateTimeCalendar extends LitElement {
   });
 
   private _focusLeave = new GUIFocusLeaveController(this, {
-    attach: 'manual',
-    defer: 'raf',
     onLeave: () => {
       this.dispatchEvent(new CustomEvent('blur', { bubbles: true, composed: true }));
     },
