+++ libs/gui/components/src/lib/controllers/focus-leave.controller.ts
@@ -5,44 +5,11 @@ export type GUIFocusLeaveHost = ReactiveControllerHost & HTMLElement;
 export interface GUIFocusLeaveControllerOptions {
   /** Called when focus has effectively left the watched subtree. */
   onLeave(): void;
-  /**
-   * Containment test applied to both the focusout `relatedTarget` and the
-   * deferred `document.activeElement` re-check. Defaults to
-   * `(el) => !!el && host.contains(el)`, which is what the pickers and
-   * calendars use today (light DOM, so `contains` sees the whole widget).
-   */
-  isInside?(el: Element | null): boolean;
-  /**
-   * How to react once the focusout target is outside:
-   * - 'raf' (default): defer via requestAnimationFrame — storing the rAF id,
-   *   cancelling any pending one, and re-checking `document.activeElement`
-   *   containment before calling `onLeave`. This is the behavior of the
-   *   date/date-time/range pickers and the calendars.
-   * - 'immediate': call `onLeave` synchronously, with no re-check. This is
-   *   the behavior of time-picker, whose focusout close has no rAF deferral.
-   */
-  defer?: 'raf' | 'immediate';
-  /**
-   * - 'host' (default): attach a `focusout` listener on the host element in
-   *   hostConnected and remove it in hostDisconnected — how the pickers wire
-   *   focusout today via connectedCallback.
-   * - 'manual': attach nothing; the owner binds `onFocusOut` in a template
-   *   (`@focusout=${ctrl.onFocusOut}`, like abstract-calendar's wrapper div)
-   *   or forwards events to `handleFocusOut` itself.
-   */
-  attach?: 'host' | 'manual';
 }
 
 /**
  * Detects "focus left the widget" with the exact semantics duplicated across
- * the six picker components and abstract-calendar:
- *
- * 1. If the focusout `relatedTarget` is still inside, do nothing.
- * 2. Otherwise defer one animation frame (cancelling any pending deferral)
- *    and re-check `document.activeElement`; only call `onLeave` if focus is
- *    really outside by then. The deferral absorbs transient focus moves
- *    (e.g. month navigation re-rendering buttons).
- * 3. A pending deferral is cancelled on hostDisconnected.
+ * the six picker components:
  */
 export class GUIFocusLeaveController implements ReactiveController {
   private host: GUIFocusLeaveHost;
@@ -66,11 +33,6 @@ export class GUIFocusLeaveController implements ReactiveController {
       return;
     }
 
-    if (this.options.defer === 'immediate') {
-      this.options.onLeave();
-      return;
-    }
-
     if (this._rafId !== undefined) {
       cancelAnimationFrame(this._rafId);
     }
@@ -83,7 +45,7 @@ export class GUIFocusLeaveController implements ReactiveController {
   }
 
   /** Cancels a pending deferred leave check, if any. */
-  cancelPendingLeave() {
+  private cancelPendingLeave() {
     if (this._rafId !== undefined) {
       cancelAnimationFrame(this._rafId);
       this._rafId = undefined;
@@ -91,22 +53,10 @@ export class GUIFocusLeaveController implements ReactiveController {
   }
 
   private isInside(el: Element | null): boolean {
-    if (this.options.isInside) {
-      return this.options.isInside(el);
-    }
     return !!el && this.host.contains(el);
   }
 
-  hostConnected() {
-    if ((this.options.attach ?? 'host') === 'host') {
-      this.host.addEventListener('focusout', this.onFocusOut);
-    }
-  }
-
   hostDisconnected() {
-    if ((this.options.attach ?? 'host') === 'host') {
-      this.host.removeEventListener('focusout', this.onFocusOut);
-    }
     this.cancelPendingLeave();
   }
 }
